requirements.txt
# Core dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Streamlit UI
streamlit>=1.28.0
requests>=2.31.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0

# Embeddings
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
open-clip-torch>=2.24.0

# Vector stores
faiss-cpu>=1.7.4; sys_platform != "linux" or platform_machine != "x86_64"
faiss-cpu>=1.7.4; sys_platform == "linux" and platform_machine == "x86_64"
# For GPU support, use: faiss-gpu>=1.7.4
qdrant-client>=1.7.0

# LLM models
llama-cpp-python>=0.2.0
ollama>=0.1.0

# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0

# Evaluation
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation (optional)
mkdocs>=1.5.0
mkdocs-material>=9.4.0

